+++ inst/extdata/testdata/corrupt/custom_pedigree_maternity_duo_pedigree.R
@@ -1,8 +1,0 @@
-#Start the Familias package
-#It must be already downloaded with e.g.
-#install.packages(Familias)
-library(Familias)
-#Mother/child
-persons <- c("C2","C1")
-pedigree <- FamiliasPedigree(id=persons, dadid=c(NA,NA), momid=c("C1", NA),
-                         sex=c("male", "female"))